backend/pyproject.toml
[project]
name = "konflikt-simulator"
version = "0.1.0"
description = "KI-gestützte Konfliktsimulation mit LangGraph"
authors = [
    {name = "Maintainer"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "langchain-anthropic (>=1.3.0,<2.0.0)",
    "anthropic (>=0.75.0,<0.76.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
